chore(admin): convert UseCasePage to TS
