+++ packages/core/admin/admin/src/pages/App/index.js
@@ -28,7 +28,7 @@ import { useEnterprise } from '../../hooks/useEnterprise';
 import { createRoute, makeUniqueRoutes } from '../../utils';
 import AuthPage from '../AuthPage';
 import { NotFoundPage } from '../NotFoundPage';
-import UseCasePage from '../UseCasePage';
+import { UseCasePage } from '../UseCasePage';
 
 import { ROUTES_CE, SET_ADMIN_PERMISSIONS } from './constants';
